requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
enum-compat==0.0.3
filelock==3.8.0
future==0.18.2
huggingface-hub==0.10.1
idna==3.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
Pillow==9.2.0
psutil==5.9.3
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
SCons==4.4.0
timm==0.6.11
tokenizers==0.13.1
torch==1.13.0
torch-model-archiver==0.6.0
torchserve==0.6.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
